pyproject.toml
[tool.poetry]
name = "ldimbenchmark"
version = "0.1.12"
description = ""
authors = ["DanielHabenicht <daniel-habenicht@outlook.de>"]
readme = "README.md"

[tool.poetry.scripts]
ldimbenchmark = "ldimbenchmark.cli:cli"

[tool.poetry.dependencies]
python = "^3.8"
tqdm = "^4.64.1"
wntr = "^0.5.0"
pyyaml = "^6.0"
scikit-learn = "^1.1.3"
docker = "^6.0.1"
tabulate = "^0.9.0"
pydantic = "^1.10.2"
big-o = "^0.10.2"
click = "^8.1.3"

[tool.poetry.group.dev.dependencies]
tox = "^3.27.1"
mkdocs-material = "^8.5.10"
black = "^22.10.0"
pytest = "^7.2.0"
syrupy = "^3.0.5"
pytest-xdist = "^3.1.0"
ipykernel = "^6.19.2"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}
mkdocs-gen-files = "^0.4.0"
mkdocs-literate-nav = "^0.5.0"
mkdocs-section-index = "^0.3.4"
mkdocs-jupyter = "^0.22.0"


[tool.poetry.group.ci.dependencies]
tox-gh-actions = "^2.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = True
envlist = py38,py39,py310

[gh-actions]
python =
    3.8: py38
    3.9: py39
    3.10: py310

[testenv]
deps =
    black
    pytest
commands =
    black --check ldimbenchmark
    pytest .
"""